Moved components from app/helpers into app/components.
